+++ pkg/azureclients/armclient/azure_armclient.go
@@ -93,6 +93,20 @@ func sender() autorest.Sender {
 		}
 		j, _ := cookiejar.New(nil)
 		defaultSenders.sender = &http.Client{Jar: j, Transport: roundTripper}
+
+		// In go-autorest SDK https://github.com/Azure/go-autorest/blob/master/autorest/sender.go#L258-L287,
+		// if ARM returns http.StatusTooManyRequests, the sender doesn't increase the retry attempt count,
+		// hence the Azure clients will keep retrying forever until it get a status code other than 429.
+		// So we explicitly removes http.StatusTooManyRequests from autorest.StatusCodesForRetry.
+		// Refer https://github.com/Azure/go-autorest/issues/398.
+		// TODO(feiskyer): Use autorest.SendDecorator to customize the retry policy when new Azure SDK is available.
+		statusCodesForRetry := make([]int, 0)
+		for _, code := range autorest.StatusCodesForRetry {
+			if code != http.StatusTooManyRequests {
+				statusCodesForRetry = append(statusCodesForRetry, code)
+			}
+		}
+		autorest.StatusCodesForRetry = statusCodesForRetry
 	})
 	return defaultSenders.sender
 }
