+++ pkg/azureclients/v2/cred/cred.go
@@ -1,0 +1,1 @@
+package cred
